Move TileInfoNormalizer to logic
